VoteForm to VoteItems
